pyproject.toml
[project]
name = "bytekart-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "cryptography>=46.0.5",
    "fastapi>=0.129.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.35",
    "uvicorn>=0.41.0",
]
